Update CONTRIBUTING.adoc (#224)
Very small fix to the CONTRIBUTING document.  The ":Z" is in the wrong position. In its current location, it is trying to pull an image called "antora/antora:Z". I believe the intent was to have the "Z" on the volume mount option to add the proper SELinux label to the volume mount. Tested this updated command successfully on Fedora 34 workstation box.